Rename test folder to tests in the runtime
